Refactoring: split GrafanaJsonSupport to multiple classes
