+++ src/main/scala/be/cetic/tsorage/hub/grafana/GrafanaService.scala
@@ -5,8 +5,10 @@ import java.time.Instant
 import akka.event.Logging
 import akka.http.scaladsl.server.{Directives, Route, StandardRoute}
 import akka.http.scaladsl.server.directives.DebuggingDirectives
+import be.cetic.tsorage.hub.grafana.grafanajsonsupport.{AnnotationObject, AnnotationRequest, AnnotationResponse,
+  DataPoints, GrafanaJsonSupport, QueryRequest, QueryResponse, SearchRequest, SearchResponse}
 
-class GrafanaService extends Directives with JsonSupport {
+class GrafanaService extends Directives with GrafanaJsonSupport {
   val database: FakeDatabase.type = FakeDatabase
 
   /**
