+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/Annotation.scala
@@ -1,0 +1,8 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * An annotation.
+ *
+ */
+final case class Annotation(name: String, enable: Boolean, datasource: String, iconColor: Option[String],
+                            query: Option[String])
