+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/AnnotationRequest.scala
@@ -1,0 +1,8 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+
+/**
+ * A request for the annotation route ("/annotations").
+ *
+ */
+final case class AnnotationRequest(annotation: Annotation)
