+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/AnnotationResponse.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A response for the annotation route ("/annotations").
+ *
+ */
+final case class AnnotationResponse(annotations: Seq[AnnotationObject])
