+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/QueryResponse.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A response for the query route ("/query").
+ *
+ */
+final case class QueryResponse(dataPointsSeq: Seq[DataPoints])
