+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/SearchRequest.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A request for the search route ("/search").
+ *
+ */
+final case class SearchRequest(target: Option[String])
