+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/SearchResponse.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A response for the search route ("/search").
+ *
+ */
+final case class SearchResponse(targets: Seq[String])
