+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/Target.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A target (metric name).
+ *
+ */
+final case class Target(target: Option[String])
