+++ src/main/scala/be/cetic/tsorage/hub/grafana/grafanajsonsupport/TimeRange.scala
@@ -1,0 +1,7 @@
+package be.cetic.tsorage.hub.grafana.grafanajsonsupport
+
+/**
+ * A range of time in ISO 8601 date format.
+ *
+ */
+final case class TimeRange(from: String, to: String)
